chnages done changed name to HostUrl
changes datatype for topcount

+++ Bikewale.Cache/BikeData/BikeVersionsCacheRepository.cs
@@ -241,7 +241,7 @@ namespace Bikewale.Cache.BikeData
         /// <param name="topcount"></param>
         /// <param name="cityId"></param>
         /// <returns>SimilarBikesForEMIEntityList</returns>
-        public IEnumerable<SimilarBikesForEMIEntity> GetSimilarBikesForEMI(int versionId, short topcount, int cityId)
+        public IEnumerable<SimilarBikesForEMIEntity> GetSimilarBikesForEMI(int versionId, byte topcount, int cityId)
         {
             IEnumerable<SimilarBikesForEMIEntity> objBikes = null;
